Move phone number format/parse into ts/types/PhoneNumber
