+++ ts/types/PhoneNumber.ts
@@ -1,6 +1,6 @@
-import { instance, PhoneNumberFormat } from './libphonenumberInstance';
+import { instance, PhoneNumberFormat } from '../util/libphonenumberInstance';
 
-export function formatPhoneNumber(
+export function format(
   phoneNumber: string,
   options: {
     ourRegionCode: string;
@@ -19,4 +19,20 @@ export function formatPhoneNumber(
   } catch (error) {
     return phoneNumber;
   }
+}
+
+export function parse(
+  phoneNumber: string,
+  options: {
+    regionCode: string;
+  }
+): string {
+  const { regionCode } = options;
+  const parsedNumber = instance.parse(phoneNumber, regionCode);
+
+  if (instance.isValidNumber(parsedNumber)) {
+    return instance.format(parsedNumber, PhoneNumberFormat.E164);
+  }
+
+  return phoneNumber;
 }
